requirements.txt
asgiref>=3.4.0
bcrypt==4.0.1
click>=7.*
cryptography>=3.0
fastapi==0.99.0
httptools>=0.2.0
jwt==1.3.1
passlib==1.7.4
pydantic==1.10.9
psycopg2-binary==2.9.6
python-multipart==0.0.6
python-dotenv==1.0.0
starlette==0.27.0
SQLAlchemy==2.0.18
typing-extensions>=3.10.0
uvicorn==0.22.0
